+++ test/content/config-loading/_config.js
@@ -1,6 +1,7 @@
 const data = {
   addOns: ['utils', 'config.js'],
   some: 'data',
+  override: 'root',
   isLevel0: true
 }