+++ test/content/config-loading/level1/_config.js
@@ -1,4 +1,5 @@
 const data = {
+  override: 'level1',
   isLevel1: true
 }