+++ test/content/config-loading/level1/level1_1/_config.js
@@ -1,4 +1,5 @@
 const data = {
+  override: 'level1_1',
   isLevel1_1: true
 }